+++ src/path_util.rs
@@ -1,3 +1,4 @@
+use std::env;
 use std::ffi::CString;
 use std::os::unix::ffi::OsStrExt;
 use std::path::{Path, PathBuf};
@@ -84,3 +85,22 @@ impl<'a, T:AsRef<[u8]>> ToCString for T {
     }
 }
 
+pub trait Expand {
+    fn expand_home(self) -> Result<PathBuf, ()>;
+}
+
+impl Expand for PathBuf {
+    fn expand_home(self) -> Result<PathBuf, ()> {
+        if !self.starts_with("~") {
+            return Ok(self);
+        }
+        let mut it = self.iter();
+        it.next();
+        if let Some(home) = env::var_os("VAGGA_USER_HOME") {
+            let home = Path::new(&home);
+            Ok(home.join(it.as_path()))
+        } else {
+            Err(())
+        }
+    }
+}